refactor: destructured and put types in types folder
