+++ src/components/contributorpage/ContributorComp.tsx
@@ -1,5 +1,5 @@
 import { getFormattedDate } from '@/lib/Date';
-import type { TCont } from '@/server/models/validation/contributor.validation';
+import type { TCont } from '@/types/server/Contributors';
 import GithubIcon from '@icons/Github';
 import Glow from '@utilities/Glow';
 import Image from 'next/image';
