+++ src/lib/Contributors.ts
@@ -1,9 +1,6 @@
 import axios from '@/config/axios';
 import octokit from '@/server/config/octokit';
-import {
-  TCont,
-  TContributor,
-} from '@/server/models/validation/contributor.validation';
+import { TCont, TContributor } from '@/types/server/Contributors';
 import fs from 'fs';
 import matter from 'gray-matter';
 import path from 'path';
