+++ src/pages/api/v1/contributors/total.ts
@@ -1,6 +1,6 @@
 import { getTotalContributors } from '@/server/controllers/contributors';
 import connectDB from '@/server/db/connectDB';
-import { NextApiHandler } from 'next';
+import type { NextApiHandler } from 'next';
 
 // end-point: origin/api/v1/contributors
 const handler: NextApiHandler = async (req, res) => {
