+++ src/server/models/validation/contributor.validation.ts
@@ -18,14 +18,3 @@ const ContV = z.object({
   following: z.number().default(0),
 });
 export default ContV;
-export type TContributor = z.infer<typeof ContV>;
-export type TCont = Pick<
-  TContributor,
-  | 'avatar_url'
-  | 'gh_username'
-  | 'name'
-  | 'bio'
-  | 'occupation'
-  | 'createdAt'
-  | 'followers'
->;
